requirements.txt
fastapi==0.92.0
uvicorn==0.20.0
cassandra-driver==3.25.0
python-dotenv==0.21.1
email-validator==1.3.1
argon2-cffi==21.3.0
pytest==7.2.1
jinja2==3.1.2
python-multipart==0.0.6
python-jose[cryptography]